source_code/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.1
Django==5.1.1
django-decouple==2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
numpy==2.1.2
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
pytesseract==0.3.13
python-decouple==3.8
sqlparse==0.5.1
tzdata==2024.2
